prediction fixed, map generated, conca.. need fixin

+++ ltae/dataset.py
@@ -68,9 +68,8 @@ class SITSData(data.Dataset):
         self.X = self.X_[idx]
         self.y = self.y_[idx]
 
-        self.X = np.array(self.X, dtype = float)
-        self.y = np.array(self.y, dtype = int)
-        
+        self.X = np.array(self.X, dtype = "float16")
+        self.y = np.array(self.y, dtype = "int8")
         self.X = self.X.reshape(int(self.X.shape[0]/n_channel), n_channel)
 
         # transform
